requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
coverage==4.5.1
Django==2.1.2
djangorestframework==3.8.2
flower==0.9.2
kombu==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
redis==2.10.6
six==1.11.0
tornado==5.1.1
vine==1.1.4
